Feature/ucp 335 create business units restrictions (#138) (#139)

+++ src/sga/business-unit/application/delete-classroom/delete-classroom.handler.ts
@@ -2,6 +2,7 @@ import { CommandHandler } from '#shared/domain/bus/command.handler';
 import { ClassroomRepository } from '#business-unit/domain/repository/classroom.repository';
 import { DeleteClassroomCommand } from '#business-unit/application/delete-classroom/delete-classroom.command';
 import { ClassroomGetter } from '#business-unit/domain/service/classroom-getter.service';
+import { BusinessUnitNotFoundException } from '#shared/domain/exception/business-unit/business-unit-not-found.exception';
 
 export class DeleteClassroomHandler implements CommandHandler {
   constructor(
@@ -11,6 +12,17 @@ export class DeleteClassroomHandler implements CommandHandler {
 
   async handle(command: DeleteClassroomCommand): Promise<void> {
     const classroom = await this.classroomGetter.get(command.classroomId);
+    const businessUnits = classroom.examinationCenter.businessUnits;
+
+    if (
+      businessUnits.length > 0 &&
+      !businessUnits.find((bu) =>
+        command.adminUserBusinessUnits.includes(bu.id),
+      )
+    ) {
+      throw new BusinessUnitNotFoundException();
+    }
+    classroom.examinationCenter.businessUnits;
 
     await this.repository.delete(classroom);
   }
